requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-import-export==2.8.0
django-simple-history==3.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
et-xmlfile==1.1.0
fontawesomefree==6.1.1
fonttools==4.37.1
frozenlist==1.3.1
future==0.18.2
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.5.1
multidict==6.0.2
mypy-extensions==0.4.3
mysqlclient==2.1.1
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.1
pathspec==0.10.1
Pillow==9.1.0
platformdirs==2.5.2
psycopg2==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyVimeo==1.1.0
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tinydb==4.7.0
tomli==2.0.1
tuspy==1.0.0
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.1
zipp==3.8.1
